backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
passlib
bcrypt==3.2.2